extract struct array types into separate files
